Scripts/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.2
openpyxl==3.1.2
matplotlib==3.8.4
seaborn==0.13.2
numpy==1.26.4